+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Initial release of convenience types in the System.ClientModel namespace, including `ClientResult<T>`, `KeyCredential`, and `ClientResultException`.
+
 ### Breaking Changes
 
 ### Bugs Fixed